Feat/resume refactor

+++ src/main/java/org/nexusscode/backend/resume/service/ResumeService.java
@@ -7,11 +7,10 @@ import org.nexusscode.backend.application.service.ApplicationService;
 import org.nexusscode.backend.global.exception.CustomException;
 import org.nexusscode.backend.global.exception.ErrorCode;
 import org.nexusscode.backend.resume.domain.Resume;
-import org.nexusscode.backend.resume.dto.ResumeItemRequestDto;
-import org.nexusscode.backend.resume.dto.ResumeItemResponseDto;
-import org.nexusscode.backend.resume.dto.ResumeRequestDto;
 import org.nexusscode.backend.resume.dto.ResumeResponseDto;
 import org.nexusscode.backend.resume.repository.ResumeRepository;
+import org.nexusscode.backend.user.domain.User;
+import org.nexusscode.backend.user.service.UserService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,43 +20,58 @@ public class ResumeService {
 
     private final ApplicationService applicationService;
     private final ResumeRepository resumeRepository;
+    private final UserService userService;
 
     @Transactional
-    public ResumeResponseDto createResume(Long applicationId, ResumeRequestDto resumeRequestDto) {
+    public ResumeResponseDto createResume(Long userId,Long applicationId) {
+        User user = userService.findById(userId);
         JobApplication application = applicationService.findById(applicationId);
+
         Resume resume = Resume.builder()
             .application(application)
-            .title(resumeRequestDto.getTitle())
+            .user(user)
             .build();
         resumeRepository.save(resume);
         return new ResumeResponseDto(resume);
     }
 
-    public List<ResumeResponseDto> getAllResumes(Long applicationId) {
+    public ResumeResponseDto getResume(Long userId, Long applicationId) {
+        User user = userService.findById(userId);
         JobApplication application = applicationService.findById(applicationId);
-        List<Resume> resumeList = resumeRepository.findAllByApplication(application);
+        if(application.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_APPLICATION);
+        }
+        Resume resume = resumeRepository.findByApplication(application);
 
-        return resumeList.stream().map(ResumeResponseDto::new).toList();
+        return new ResumeResponseDto(resume);
     }
 
-    @Transactional
+   /* @Transactional
     public ResumeResponseDto updateResume(Long resumeId, ResumeRequestDto resumeRequestDto) {
         Resume resume = findById(resumeId);
         resume.updateResume(resumeRequestDto.getTitle());
         resumeRepository.save(resume);
 
         return new ResumeResponseDto(resume);
-    }
+    }*/
 
     @Transactional
-    public void deleteResume(Long resumeId) {
+    public void deleteResume(Long userId,Long resumeId) {
+        User user = userService.findById(userId);
         Resume resume = findById(resumeId);
+        if(resume.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_RESUME);
+        }
         resumeRepository.delete(resume);
     }
 
     @Transactional
-    public void saveResumeInArchieve(Long resumeId) {
+    public void saveResumeInArchieve(Long userId,Long resumeId) {
+        User user = userService.findById(userId);
         Resume resume = findById(resumeId);
+        if(resume.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_RESUME);
+        }
 
         if(resume.isSaved()){
             throw new CustomException(ErrorCode.ALREADY_SAVED_RESUME);
@@ -67,8 +81,12 @@ public class ResumeService {
     }
 
     @Transactional
-    public void cancelResumeFromArchieve(Long resumeId) {
+    public void cancelResumeFromArchieve(Long userId,Long resumeId) {
+        User user = userService.findById(userId);
         Resume resume = findById(resumeId);
+        if(resume.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_RESUME);
+        }
         if(!resume.isSaved()){
             throw new CustomException(ErrorCode.NOT_SAVED_RESUME);
         }
